modify directory name (ServerModule->SQLModule)
